+++ CHANGELOG.md
@@ -18,6 +18,9 @@ We use the following categories for changes:
 
 ### Added
 
+- `auth print-api-key` command to print groundcover api-key [#sc-4121]
+- `auth login` command to authenticate your accout in groundcover [#sc-4121]
+
 ### Changed
 
 ### Fixed